null check usage cleanup, using statement cleanup

+++ user.api/Common/Models/Data/Base/BaseEntity.cs
@@ -3,8 +3,8 @@
 public abstract class BaseEntity : IEntity
 {
     public Guid id { get; set; }
-    public int created_on { get; set; }
-    public int updated_on { get; set; }
+    public long created_on { get; set; }
+    public long updated_on { get; set; }
     public Guid? created_by { get; set; }
     public Guid? updated_by { get; set; }
 }
