+++ user.api/Common/Models/Data/Base/IEntity.cs
@@ -3,8 +3,8 @@
 public interface IEntity
 {
     Guid id { get; set; }
-    int created_on { get; set; }
-    int updated_on { get; set; }
+    long created_on { get; set; }
+    long updated_on { get; set; }
     Guid? created_by { get; set; }
     Guid? updated_by { get; set; }
 }
